+++ src/protocol/filter.rs
@@ -122,8 +122,8 @@ impl SingleLetterTag {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Object)]
-#[uniffi::export(Debug, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Object)]
+#[uniffi::export(Debug, Eq, Hash)]
 pub struct Filter {
     inner: nostr::Filter,
 }